Feature/encryption realm

+++ Quotes/Business Logic Layer/Realm/RealmService.swift
@@ -10,11 +10,39 @@ protocol RealmServicProtocol {
     func removeAllQuotesInCategory(category: String) -> Bool
 }
 
-final class RealmService: RealmServicProtocol {
+final class RealmService {
+        
+    static let encryptionKey: Data = {
+        if KeychainService.shared.isEncryptionKeyExist {
+            if let existingKey = KeychainService.shared.getEncryptionKey() {
+                return existingKey
+            }
+        } else {
+            var key = Data(count: 64)
+            _ = key.withUnsafeMutableBytes { (pointer: UnsafeMutableRawBufferPointer) in
+                SecRandomCopyBytes(kSecRandomDefault, 64, pointer.baseAddress!)
+            }
+            KeychainService.shared.saveEncryptionKey(key)
+            return key
+        }
+        fatalError("Problem with get or create encryptionKey")
+    }()
+
+    
+    private var config: Realm.Configuration?
+    
+    init() {
+        self.config = Realm.Configuration(encryptionKey: Self.encryptionKey)
+    }
+}
+
+extension RealmService: RealmServicProtocol {
     
     func saveQuote(_ quote: Quote) -> Bool {
+        guard let config = config else { return false }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             
             let handler: () -> Void = {
                 if let existingQuote = realm.object(ofType: QuoteRealmModel.self, forPrimaryKey: quote.id) {
@@ -45,8 +73,11 @@ final class RealmService: RealmServicProtocol {
     }
     
     func fetchQuotes() -> [Quote] {
+        
+        guard let config = config else { return [] }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             
             let objects = realm.objects(QuoteRealmModel.self)
             
@@ -61,8 +92,11 @@ final class RealmService: RealmServicProtocol {
     }
     
     func deleteQuote(usingID id: String) -> Bool {
+        
+        guard let config = config else { return false }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             
             let objects = realm.objects(QuoteRealmModel.self).filter{ $0.id == id }
             let handler: () -> Void = {
@@ -84,8 +118,11 @@ final class RealmService: RealmServicProtocol {
     }
     
     func removeAllQuotes() -> Bool {
+        
+        guard let config = config else { return false }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             
             if realm.isInWriteTransaction {
                 realm.deleteAll()
@@ -102,8 +139,11 @@ final class RealmService: RealmServicProtocol {
     }
     
     func fetchCategories() -> [String] {
+        
+        guard let config = config else { return [] }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             let categories = realm.objects(QuoteRealmModel.self).compactMap { $0.category }
             return Array(categories)
         } catch {
@@ -113,8 +153,11 @@ final class RealmService: RealmServicProtocol {
     }
     
     func removeAllQuotesInCategory(category: String) -> Bool {
+        
+        guard let config = config else { return false }
+        
         do {
-            let realm = try Realm()
+            let realm = try Realm(configuration: config)
             
             let objects = realm.objects(QuoteRealmModel.self).filter{ $0.category == category }
             let handler: () -> Void = {
@@ -136,5 +179,4 @@ final class RealmService: RealmServicProtocol {
         }
     }
 }
-
